Consolidated satis-s3 images
Previously mojdigital/satis-s3 was based on iainmackay/satis-s3.
Now they have been consolidated into one image.
This was to allow the arguments to run.sh to be modified,
as previously they were hardcoded in iainmackay/satis-s3.
These arguments can now be specified as environment variables.

This step was taken, due to the changes to the CircleCi YML format.
The previous was volumes were mounted could not be replicated with the 2.0 format.